requirements.txt
requests
streamlit
pandas
psycopg2-binary
python-dotenv
sqlalchemy
